create route for data upload
