+++ models/Receipts.js
@@ -25,4 +25,4 @@ const ReceiptSchema = new Schema({
   comments: String,
 });
 
-module.exports = mongoose.model('rtests', ReceiptSchema);
+module.exports = mongoose.model('receipts', ReceiptSchema);
