+++ uploads/a1e1b20a7d674ec56dc0dbbb6808b2a2
@@ -1,0 +1,6 @@
+﻿late,daysLate,transactionId,poNumber,lineNumber,vendorName,vendorSite,itemNumber,description,poType,unitPrice,buyer,qtyTransacted,transactionDate,needByDate,promiseDate,releaseNum,internalOrExternal,reasonCode,comments
+N,0,1005,PK03208,1,ARROW ELECTRONIC INC,11488,SE26-134,SWITCH-KEY CAP,BLANKET,0.15,Herwaldt,1000,2/11/19,2/11/19,2/11/19,3,Internal,quality,
+Y,1,1006,PK16540,1,MC LOONE METAL,12447,SD43-CI6X-102,MODEL LABEL - CI6X,BLANKET,0.578,Ellis,300,6/6/19,6/6/19,6/6/19,16,External,PPAP,
+N,0,1007,PK16540,1,MC LOONE METAL,12447,SD43-CI6X-102,MODEL LABEL - CI6X,BLANKET,0.578,Ellis,150,5/9/19,5/9/19,5/9/19,15,External,PPAP,
+Y,1,1008,PK16562,1,MC LOONE METAL,12447,SD43-80,LABEL-(2 1/2 X 8 1/4),BLANKET,0.421,Ellis,200,2/19/19,2/21/19,2/21/19,7,Internal,Snag,
+N,1000,1000,PK16562,1,MC LOONE METAL,12447,SD43-80,LABEL-(2 1/2 X 8 1/4),BLANKET,0.421,Ellis,200,2/19/19,2/21/19,2/21/19,7,Internal,Snag,SNAG FOR DAYS
