+++ uploads/c588085a860b9dc7dff6222e6ad032e0
@@ -1,0 +1,10 @@
+﻿poNumber,lineNumber,vendorName,vendorSite,itemNumber,description,poType,unitPrice,buyer,qtyTransacted,transactionDate,needByDate,promiseDate,releaseNum
+2000302,1,SUPERIOR QUALITY INC,211708,VP:21170A,VP:21170A,BLANKET,140.3,Dang,4,4/25/19,3/15/19,3/15/19,6
+2000302,1,SUPERIOR QUALITY INC,211708,VP:21170A,VP:21170A,BLANKET,140.3,Dang,6,4/12/19,3/15/19,3/15/19,6
+P205067,1,TECHPRINT INC,69362,VP:69362A,VP:69362A,BLANKET,2.75,Dang,600,1/22/19,1/30/19,1/30/19,6
+2000154,1,"COXON PRECISE INDUSTRIAL CO., LTD",184762,KOH-26096-A,ENCLOSURE LEFT - BLUE,BLANKET,1.9978,DeCook,162,1/17/19,2/1/19,2/1/19,3
+2000155,1,"COXON PRECISE INDUSTRIAL CO., LTD",184762,KOH-26095-A,ENCLOSURE RIGHT - BLUE,BLANKET,2.2515,DeCook,108,5/9/19,5/24/19,5/9/19,4
+PK03208,1,ARROW ELECTRONIC INC,11488,SE26-134,SWITCH-KEY CAP,BLANKET,0.15,Herwaldt,1000,2/11/19,2/11/19,2/11/19,3
+PK16540,1,MC LOONE METAL,12447,SD43-CI6X-102,MODEL LABEL - CI6X,BLANKET,0.578,Ellis,300,6/6/19,6/6/19,6/6/19,16
+PK16540,1,MC LOONE METAL,12447,SD43-CI6X-102,MODEL LABEL - CI6X,BLANKET,0.578,Ellis,150,5/9/19,5/9/19,5/9/19,15
+PK16562,1,MC LOONE METAL,12447,SD43-80,LABEL-(2 1/2 X 8 1/4),BLANKET,0.421,Ellis,200,2/19/19,2/21/19,2/21/19,7
